refactor: 패키지 네이밍 변경(gateway -> service)
- Order 트랜잭션 어노테이션 적용
